+++ example/building/floor.py
@@ -1,10 +1,15 @@
 import cadquery as cq
 from cqterrain.building import Floor
 
-bp = Floor()
-bp.make()
-f = bp.build()
-workspace = cq.Workplane('XY')
-workspace.add(f)
+bp_floor = Floor()
+bp_floor.length = 100 
+bp_floor.width = 100
+bp_floor.height = 3 
+bp_floor.tile = None 
+bp_floor.tile_padding = 0
 
-cq.exporters.export(workspace,'stl/floor.stl')
+bp_floor.make()
+result = bp_floor.build()
+
+#show_object(floor_ex)
+cq.exporters.export(result,'stl/building_floor.stl')