+++ src/cqterrain/building/Room.py
@@ -12,7 +12,8 @@
 # See the License for the specific language governing permissions and
 
 import cadquery as cq
-from cadqueryhelper import series, shape
+from cadqueryhelper import series, shape, Base
+from typing import Callable
 from . import Floor
 from . import Wall
 from .. import Door
@@ -35,44 +36,48 @@ def _make_custom_door(wall, length, width, height, floor_height):
     return w
 
 
-class Room:
+class Room(Base):
+    '''
+    This code is doing too much work. I'm going to refactor this whole concept
+    '''
     def __init__(
     self,
-    length=120,
-    width=80,
-    height=50,
-    wall_width=3,
-    floor_height=3,
-    floor_padding=0,
-    floor_tile=None,
-    floor_tile_padding=0,
-    window_count=1,
-    style="office",
-    door_walls = [False, True, False, False],
-    window_walls = [True, True, True, True],
-    build_walls = [True, True, True, True],
-    make_custom_windows = None,
-    make_custom_door = None
+    length:float = 120,
+    width:float = 80,
+    height:float = 50,
+    wall_width:float = 3,
+    floor_height:float = 3,
+    floor_padding:float = 0,
+
+    floor_tile:cq.Workplane|None = None,
+    floor_tile_padding:float = 0,
+    window_count:int = 1,
+    style:str = "office",
+    door_walls:list[bool] = [False, True, False, False],
+    window_walls:list[bool] = [True, False, True, True],
+    build_walls:list[bool] = [True, True, True, True],
+    make_custom_windows:Callable[[cq.Workplane, float, float, float, int, float], cq.Workplane]|None = None,
+    make_custom_door:Callable[[cq.Workplane, float, float, float, float], cq.Workplane]|None = None
     ):
         # attributes
-        self.length = length
-        self.width = width
-        self.height = height
-        self.wall_width = wall_width
-        self.floor_height = floor_height
-        self.floor_padding = floor_padding
-        self.floor_tile = floor_tile
-        self.floor_tile_padding = floor_tile_padding
-        self.style = style
-        self.window_count = window_count
-        self.door_walls = door_walls
-        self.window_walls = window_walls
-        self.build_walls = build_walls
+        self.length:float = length
+        self.width:float = width
+        self.height:float = height
+        self.wall_width:float = wall_width
+        self.floor_height:float = floor_height
+        self.floor_padding:float = floor_padding
+        self.floor_tile:cq.Workplane|None = floor_tile
+        self.floor_tile_padding:float = floor_tile_padding
+        self.style:str = style
+        self.window_count:int = window_count
+        self.door_walls:list[bool] = door_walls
+        self.window_walls:list[bool] = window_walls
+        self.build_walls:list[bool] = build_walls
 
         # post make
-        self.floor = None
-        self.walls = []
-        self.doors = []
+        self.floor:Base|None = None
+        self.walls:list[cq.Workplane] = []
+        self.doors:list[cq.Workplane] = []
 
         self.window = {}
         self.window['padding'] = 1
@@ -101,22 +106,34 @@ class Room:
         self.r_length = floor_bp.length - padding
         floor_bp.make()
         return floor_bp
-        #r_floor = floor_bp.build()
-        #return r_floor
 
-    def __make_wall(self, length, width, height, door_wall, window_wall):
-        b_wall = Wall(length, width, height)
-        b_wall.make()
-        w = b_wall.build()
-        #print('wall built', dir(w))
-        self.w_height = b_wall.height
-        self.w_width = b_wall.width
-
+    def __make_wall(
+            self, 
+            length:float, 
+            width:float, 
+            height:float, 
+            door_wall:bool, 
+            window_wall:bool
+        ):
+        # this method is doing too much work
+        # I don't like this, I want the blueprints to be parameters.
+        bp_wall:Base = Wall(length, width, height)
+        bp_wall.make()
+        wall:cq.Workplane = bp_wall.build()
+        self.w_height = bp_wall.height
+        self.w_width = bp_wall.width
 
         if self.make_custom_windows != None and window_wall:
-            w = self.make_custom_windows(w, self.window['length'], width, self.window['height'], self.window_count, self.window['padding'])
+            wall = self.make_custom_windows(
+                wall, 
+                self.window['length'], 
+                width, 
+                self.window['height'], 
+                self.window_count, 
+                self.window['padding']
+            )
         elif self.style == "office" and window_wall:
-            w = _make_custom_windows(w, self.window['length'], width, self.window['height'], self.window_count, self.window['padding'])
+            wall = _make_custom_windows(wall, self.window['length'], width, self.window['height'], self.window_count, self.window['padding'])
         elif self.style == "arch" and window_wall:
             window_cutout = shape.arch_pointed(length=12, width=width, height=22, inner_height=10)
             window_series = series(window_cutout, self.window_count, length_offset = self.window['padding'])
@@ -125,31 +142,31 @@ class Room:
             window_cutout2 = shape.arch_pointed(length=10, width=width+2, height=20, inner_height=10)
             window = window_ridge.cut(window_cutout2)
             window_series2 = series(window, self.window_count, length_offset = self.window['padding'])
-            w = w.cut(window_series).add(window_series2)
-
+            wall = wall.cut(window_series).add(window_series2)
 
         if door_wall and self.make_custom_door:
-            w = self.make_custom_door(w, self.door['length'], self.door['width'], self.door['height'], self.floor_height)
+            w = self.make_custom_door(wall, self.door['length'], self.door['width'], self.door['height'], self.floor_height)
         elif door_wall:
             #print('attempt to make door')
             #zero wall
-            w = w.translate((0,0,height/2))
+            wall = wall.translate((0,0,height/2))
 
             # door logic
-            door_bb = Door(**self.door)
-            door_bb.make()
-            door_outline = door_bb.outline.translate((0,0,(door_bb.height/2)+self.floor_height))
-            door = door_bb.build().translate((0,0,(door_bb.height/2)+self.floor_height))
-            w = w.cut(door_outline).union(door)
+            bp_door = Door(**self.door)
+            bp_door.make()
+            door_outline = bp_door.outline.translate((0,0,(bp_door.height/2)+self.floor_height)) #type: ignore
+            door = bp_door.build().translate((0,0,(bp_door.height/2)+self.floor_height))
+            wall = wall.cut(door_outline).union(door)
 
             #center all
-            w = w.translate((0,0,-1*(height/2)))
-        return w
+            wall = wall.translate((0,0,-1*(height/2)))
+        return wall
 
     def make(self):
         # make floor
         self.floor = self.__make_floor()
 
+        # this code shouldn't be doing the rotations for placement, that's a build problem
         # make walls along the x axis
         w1  = self.__make_wall(length=self.r_length, width=self.wall_width, height=self.height, door_wall = self.door_walls[0], window_wall = self.window_walls[0])
         w1_rotated = w1.rotate((0, 0, 1), (0, 0, 0), 180)
@@ -169,9 +186,54 @@ class Room:
         self.walls.append(w4_rotated)
 
     def build(self):
+        scene = cq.Workplane("XY")
+
+        if self.floor:
+            floor = self.floor.build()
+            scene = scene.union(floor)
+            
+        if self.build_walls[0]:
+            scene = scene.union(self.walls[0].translate((
+                0, 
+                (self.r_width /2) - (self.w_width /2), 
+                (self.w_height /2)-(self.r_height/2)))
+            )
+
+        if self.build_walls[1]:
+            scene = scene.union(self.walls[1].translate((
+                0, 
+                -1*((self.r_width /2) - (self.w_width /2)), 
+                (self.w_height /2)-(self.r_height/2)))
+            )
+
+        if self.build_walls[2]:
+            scene = scene.union(self.walls[2].translate((
+                (self.r_length /2) - (self.w_width /2), 
+                0, 
+                (self.w_height /2)-(self.r_height/2)))
+            )
+
+        if self.build_walls[3]:
+            scene = scene.union(self.walls[3].translate((
+                -1*((self.r_length /2) - (self.w_width /2)), 
+                0, 
+                (self.w_height /2)-(self.r_height/2)))
+            )
+
+        # zero out height
+        scene = scene.translate((0,0, self.floor_height/2))
+        scene = scene.translate((0,0, -1*(self.height/2)))
+
+        return scene
+    
+
+    def build_assembly(self):
         room_assembly = cq.Assembly()
-        floor = self.floor.build()
-        room_assembly.add(floor, name="floor")
+
+        if self.floor:
+            floor = self.floor.build()
+            room_assembly.add(floor, name="floor")
+            
         if self.build_walls[0]:
             room_assembly.add(self.walls[0], name="wall1", loc=cq.Location(cq.Vector(0, (self.r_width /2) - (self.w_width /2), (self.w_height /2)-(self.r_height/2))))
 
@@ -183,10 +245,5 @@ class Room:
 
         if self.build_walls[3]:
             room_assembly.add(self.walls[3], name="wall4", loc=cq.Location(cq.Vector(-1*((self.r_length /2) - (self.w_width /2)), 0, (self.w_height /2)-(self.r_height/2))))
-        comp_room = room_assembly.toCompound()
 
-        # zero out height
-        comp_room = comp_room.translate((0,0, self.floor_height/2))
-        comp_room = comp_room.translate((0,0, -1*(self.height/2)))
-
-        return comp_room
+        return room_assembly